Consolidate SafeFetch/WebFetch tests into one file, one class each
The per-method test folders (tests/.../utils/safe_fetch/ and .../web_fetch/)
were the outlier layout inherited from the original WebFetch test commit; the
rest of the repo uses one test file per module (e.g. test_pdf_utils.py). Fold
them into a single test_safe_fetch.py (class TestSafeFetch) and test_web_fetch.py
(class TestWebFetch).

- safe_fetch: 8 files -> 1 file / 1 class; every test method and the six private
  _call helpers are subject-prefixed to avoid collisions in the single class
  (two test names and the _call helpers previously overlapped across files).
  helpers.py mock factories are inlined at module level.
- web_fetch: 2 files -> 1 file / 1 class; the two setUp methods are merged.
- No behavior change: same 94 tests (78 + 16), all passing; bodies preserved
  verbatim. ruff + pylint clean (10.00/10).
