--- tests/neuro_san_studio/coded_tools/utils/safe_fetch/test_get_content_type.py
@@ -1,204 +1,0 @@
-# Copyright © 2025-2026 Cognizant Technology Solutions Corp, www.cognizant.com.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# END COPYRIGHT
-
-import asyncio
-from unittest import TestCase
-from unittest.mock import AsyncMock
-from unittest.mock import MagicMock
-from unittest.mock import patch
-
-from aiohttp import ClientError
-from aiohttp import ClientResponseError
-
-from neuro_san_studio.coded_tools.utils.safe_fetch import MAX_RESPONSE_BYTES
-from neuro_san_studio.coded_tools.utils.safe_fetch import SafeFetch
-from tests.neuro_san_studio.coded_tools.utils.safe_fetch.helpers import make_head_session
-from tests.neuro_san_studio.coded_tools.utils.safe_fetch.helpers import make_response_error
-
-MODULE = "neuro_san_studio.coded_tools.utils.safe_fetch"
-
-
-class TestGetContentType(TestCase):
-    """Unit tests for SafeFetch.get_content_type."""
-
-    def test_head_success_returns_content_type(self):
-        """Tests that a successful HEAD response returns the Content-Type header value with no prefetched body."""
-        session, _ = make_head_session(status=200, content_type="text/html; charset=utf-8")
-        content_type, body = asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertEqual(content_type, "text/html; charset=utf-8")
-        self.assertIsNone(body)
-
-    def test_head_405_falls_back_to_get_pdf(self):
-        """Tests that a 405 HEAD response falls back to GET and returns content type without body for PDF."""
-        session, _ = make_head_session(status=405)
-        get_response = MagicMock()
-        get_response.status = 200
-        get_response.headers = {"Content-Type": "application/pdf"}
-        get_response.raise_for_status = MagicMock()
-        get_cm = MagicMock()
-        get_cm.__aenter__ = AsyncMock(return_value=get_response)
-        get_cm.__aexit__ = AsyncMock(return_value=False)
-        session.get = MagicMock(return_value=get_cm)
-
-        content_type, body = asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertEqual(content_type, "application/pdf")
-        self.assertIsNone(body)
-        session.get.assert_called_once()
-
-    def test_head_405_falls_back_to_get_text_returns_body(self):
-        """Tests that a 405 HEAD response falls back to GET and returns the body for text content types."""
-        session, _ = make_head_session(status=405)
-        get_response = MagicMock()
-        get_response.status = 200
-        get_response.headers = {"Content-Type": "text/html"}
-        get_response.charset = "utf-8"
-        get_response.raise_for_status = MagicMock()
-
-        async def iter_chunked(_chunk_size):
-            yield b"<html>Hello</html>"
-
-        get_response.content.iter_chunked = iter_chunked
-        get_cm = MagicMock()
-        get_cm.__aenter__ = AsyncMock(return_value=get_response)
-        get_cm.__aexit__ = AsyncMock(return_value=False)
-        session.get = MagicMock(return_value=get_cm)
-
-        content_type, body = asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertEqual(content_type, "text/html")
-        self.assertEqual(body, "<html>Hello</html>")
-
-    def test_head_405_non_text_type_skips_body_read(self):
-        """Tests that a 405 fallback GET does not read the body for non-text content types.
-
-        A binary/unsupported type is returned with body None so the caller rejects it
-        without downloading a body that would only be discarded.
-        """
-        session, _ = make_head_session(status=405)
-        get_response = MagicMock()
-        get_response.status = 200
-        get_response.headers = {"Content-Type": "image/png"}
-        get_response.raise_for_status = MagicMock()
-        get_response.text = AsyncMock(return_value="binary-should-not-be-read")
-        get_cm = MagicMock()
-        get_cm.__aenter__ = AsyncMock(return_value=get_response)
-        get_cm.__aexit__ = AsyncMock(return_value=False)
-        session.get = MagicMock(return_value=get_cm)
-
-        content_type, body = asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertEqual(content_type, "image/png")
-        self.assertIsNone(body)
-        get_response.text.assert_not_awaited()
-
-    def test_head_405_text_body_over_limit_raises(self):
-        """Tests that the 405 text prefetch enforces the streamed byte cap, not just Content-Length.
-
-        No Content-Length header is set, so only the running byte count on the
-        streamed body can catch an oversized response (the server-lies case).
-        """
-        session, _ = make_head_session(status=405)
-        get_response = MagicMock()
-        get_response.status = 200
-        get_response.headers = {"Content-Type": "text/html"}
-        get_response.charset = "utf-8"
-        get_response.raise_for_status = MagicMock()
-
-        async def iter_chunked(_chunk_size):
-            yield b"x" * 50
-
-        get_response.content.iter_chunked = iter_chunked
-        get_cm = MagicMock()
-        get_cm.__aenter__ = AsyncMock(return_value=get_response)
-        get_cm.__aexit__ = AsyncMock(return_value=False)
-        session.get = MagicMock(return_value=get_cm)
-
-        with patch(f"{MODULE}.MAX_RESPONSE_BYTES", 10):
-            with self.assertRaises(ValueError) as ctx:
-                asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("response_too_large", str(ctx.exception))
-
-    def test_non_2xx_raises_with_url_not_accessible_prefix(self):
-        """Tests that a non-2xx HTTP error raises ClientResponseError with url_not_accessible prefix."""
-        exc = make_response_error(404)
-        session, _ = make_head_session(status=404, raise_for_status_exc=exc)
-        with self.assertRaises(ClientResponseError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("url_not_accessible", ctx.exception.message)
-        self.assertEqual(ctx.exception.status, 404)
-
-    def test_429_raises_with_too_many_requests_prefix(self):
-        """Tests that a 429 response raises ClientResponseError with too_many_requests prefix."""
-        exc = make_response_error(429)
-        session, _ = make_head_session(status=429, raise_for_status_exc=exc)
-        with self.assertRaises(ClientResponseError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("too_many_requests", ctx.exception.message)
-
-    def test_connection_error_raises_with_url_not_accessible_prefix(self):
-        """Tests that a connection error raises ClientError with url_not_accessible prefix."""
-        head_cm = MagicMock()
-        head_cm.__aenter__ = AsyncMock(side_effect=ClientError("DNS failure"))
-        head_cm.__aexit__ = AsyncMock(return_value=False)
-        session = MagicMock()
-        session.head = MagicMock(return_value=head_cm)
-
-        with self.assertRaises(ClientError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("url_not_accessible", str(ctx.exception))
-
-    def test_timeout_raises_with_url_not_accessible_prefix(self):
-        """Tests that a request timeout raises ClientError with url_not_accessible prefix."""
-        head_cm = MagicMock()
-        head_cm.__aenter__ = AsyncMock(side_effect=asyncio.TimeoutError())
-        head_cm.__aexit__ = AsyncMock(return_value=False)
-        session = MagicMock()
-        session.head = MagicMock(return_value=head_cm)
-
-        with self.assertRaises(ClientError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("url_not_accessible", str(ctx.exception))
-
-    def test_content_length_over_limit_raises_response_too_large(self):
-        """Tests that a Content-Length header exceeding the limit raises ValueError with response_too_large."""
-        session, _ = make_head_session(status=200, content_type="text/html", content_length=MAX_RESPONSE_BYTES + 1)
-        with self.assertRaises(ValueError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        self.assertIn("response_too_large", str(ctx.exception))
-
-    def test_head_redirect_raises_url_not_allowed(self):
-        """Tests that a 3xx HEAD response raises ValueError containing url_not_allowed and the Location URL."""
-        session, _ = make_head_session(status=301, extra_headers={"Location": "http://other.com/"})
-        with self.assertRaises(ValueError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        error = str(ctx.exception)
-        self.assertIn("url_not_allowed", error)
-        self.assertIn("http://other.com/", error)
-
-    def test_405_get_redirect_raises_url_not_allowed(self):
-        """Tests that a 405 HEAD + 3xx GET raises ValueError with url_not_allowed and the Location URL."""
-        session, _ = make_head_session(status=405)
-        get_response = MagicMock()
-        get_response.status = 302
-        get_response.headers = {"Location": "http://other.com/"}
-        get_cm = MagicMock()
-        get_cm.__aenter__ = AsyncMock(return_value=get_response)
-        get_cm.__aexit__ = AsyncMock(return_value=False)
-        session.get = MagicMock(return_value=get_cm)
-
-        with self.assertRaises(ValueError) as ctx:
-            asyncio.run(SafeFetch.get_content_type("http://example.com", session))
-        error = str(ctx.exception)
-        self.assertIn("url_not_allowed", error)
-        self.assertIn("http://other.com/", error)
